fix a warning in msvc6
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@7614 d7cf8633-e32d-0410-b094-e92efae38249

+++ cl_screen.c
@@ -1171,7 +1171,7 @@ Cr = R *  .500 + G * -.419 + B * -.0813 + 128.;
 		SCR_CaptureVideo_RIFF_Push("vprp", NULL);
 		SCR_CaptureVideo_RIFF_Write32(0); // VideoFormatToken
 		SCR_CaptureVideo_RIFF_Write32(0); // VideoStandard
-		SCR_CaptureVideo_RIFF_Write32(cls.capturevideo.framerate); // dwVerticalRefreshRate (bogus)
+		SCR_CaptureVideo_RIFF_Write32((int)cls.capturevideo.framerate); // dwVerticalRefreshRate (bogus)
 		SCR_CaptureVideo_RIFF_Write32(width); // dwHTotalInT
 		SCR_CaptureVideo_RIFF_Write32(height); // dwVTotalInLines
 		FindFraction(aspect, &n, &d, 1000);
